ChopFast/requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.3
Django==3.2.5
django-rest-framework==0.1.0
djangorestframework==3.12.4
idna==3.2
paystack==1.5.0
paystackapi==2.0.0
Pillow==8.3.1
PyJWT==1.7.1
python-decouple==3.4
pytz==2021.1
requests==2.26.0
simplejson==3.17.3
six==1.16.0
sqlparse==0.4.1
twilio==6.62.0
urllib3==1.26.6
